requirements.txt
streamlit
pdfplumber
python-docx
reportlab
